+++ client/src/pages/Balec.tsx
@@ -1,5 +1,5 @@
 function Balec() {
-    return <h1>Balec on t'a dit !!!!</h1>
+	return <h1>Balec on t'a dit !!!!</h1>;
 }
 
-export default Balec
+export default Balec;
