+++ client/src/pages/MaisonDesVoyagesDeMerde.tsx.tsx
@@ -4,12 +4,13 @@ import Loader from "../components/Loader";
 interface destination {
 	id: number;
 	name: string;
+	category: string;
 	description: string;
 	sound: string;
 	image: string;
 }
 
-function MaisonDesVoyagesDeMerde() {
+function Home() {
 	const [destinations, setDestination] = useState<destination[] | []>([]);
 	useEffect(() => {
 		fetch("http://localhost:3310/api/destinations")
@@ -17,9 +18,33 @@ function MaisonDesVoyagesDeMerde() {
 			.then((data) => setDestination(data));
 	}, []);
 
+	const handleChange = (e: string) => {
+		window.location.href = `/${e}`;
+	};
+
 	const random = Math.floor(Math.random() * destinations.length);
 	return (
 		<>
+			<nav>
+				<label>
+					<select value={""} onChange={(e) => handleChange(e.target.value)}>
+						<option value="">Choisissez votre destination</option>
+
+						<option
+							value="serieux-faut-vraiment-donner-une-url-a-toutes-les-pages"
+							key="1"
+						>
+							Voyage interstellaire
+						</option>
+						<option value="jai-oublie-de-fermer-le-frigo-ce-matin" key="2">
+							Voyage insolite
+						</option>
+						<option value="merde-et-dire-que-cest-pas-la-derniere-page" key="3">
+							Voyage temporel
+						</option>
+					</select>
+				</label>
+			</nav>
 			<section className="StarAligned">
 				{destinations.length === 0 ? (
 					<Loader />
@@ -27,6 +52,10 @@ function MaisonDesVoyagesDeMerde() {
 					<figure>
 						<h2>{destinations[random].name}</h2>
 						<p>{destinations[random].description}</p>
+						<img
+							src={destinations[random].image}
+							alt={destinations[random].name}
+						/>
 					</figure>
 				)}
 			</section>
@@ -34,4 +63,4 @@ function MaisonDesVoyagesDeMerde() {
 	);
 }
 
-export default MaisonDesVoyagesDeMerde;
+export default Home;
