+++ client/src/pages/NotFound.tsx
@@ -1,9 +1,11 @@
 function NotFound() {
-  return (
-    <>
-      <p>Rien à voir ici!</p>
-    </>
-  );
+	return (
+		<>
+			<main className="notfound">
+				<p>Rien à voir ici!</p>
+			</main>
+		</>
+	);
 }
 
 export default NotFound;
